Add tag support to identities. The main use case is to allow apps to tag the origin of identities when there are multiple applications sharing the same realm.
For example, if you develop apps A and B, and these apps tag their identities accordingly, then it's trivial to determine which identities came from which app. If, say, app B is discontinued, it's easy to purge the database of obsolete identities.

Tagging is entirely-freeform and can serve any purpose, however.